refactor(125/WS1.4): extract routes/tasks.rs (task-list handlers) (#125)
Second wave of the #125 server decomposition: mechanical extraction of
the `tasks` route category (category: "tasks" in src/api/mod.rs) out of
src/server/mod.rs into src/server/routes/tasks.rs.

Moved verbatim (pure relocation):
- 5 handlers: list_task_lists, create_task_list, list_tasks, add_task, update_task
- 5 DTOs:   CreateTaskListRequest, AddTaskRequest, UpdateTaskRequest,
             TaskListEntry, TaskEntry

Handlers widened to pub(in crate::server) (the floor forced by the
pub(super) re-export in routes/mod.rs — see PR #161 / E0364); route
registrations stay in src/server/mod.rs via the existing use routes::{...}
import block. The route set is byte-identical (route_set_matches_registry +
manifest_matches_registry pass; tests/snapshots/ untouched).

No behavior changes — this is the prerequisite for #153 (group-membership
enforcement on these handlers), which lands in a separate PR.
